wip: 单测测试改造完成一半
