+++ .gitignore
@@ -21,5 +21,4 @@
 .env.local
 
 # ide
-/.vscode
 /.idea
