--- test/BothLinkedList.test.js
@@ -1,13 +1,0 @@
-// import { BothLinkedList } from '../src/index';
-describe('双向链表相关测试', () => {
-  test('双向链表', () => {
-    // const bothLinkedList = new BothLinkedList();
-    // const insertRes = { data: 1, next: undefined, prev: undefined }
-    // bothLinkedList.insertHead(1);
-    // expect(bothLinkedList.getHead()).toEqual(insertRes)
-    // bothLinkedList.insertHead(2);
-    // bothLinkedList.deleteData(2);
-    // bothLinkedList.deleteData(1);
-    // expect(bothLinkedList.getHead()).toEqual(undefined)
-  });
-});
